backend/requirements.txt
django~=5.1
djangorestframework~=3.15
django-cors-headers~=4.4
celery[redis]~=5.4
redis~=5.0
psycopg[binary]~=3.2
langchain~=1.2
langchain-openai~=0.3
langchain-core~=1.2
pydantic~=2.9
